Migrando ActivityReport a frontend/src
Este cambio mueve \OmegaUp\ActivityReport al directorio que puede ser
autoloaded.
