+++ frontend/server/controllers/ContestController.php
@@ -1,6 +1,5 @@
 <?php
 
-require_once 'libs/ActivityReport.php';
 require_once 'libs/PrivacyStatement.php';
 require_once 'libs/dao/Contests.dao.php';
 
@@ -804,7 +803,10 @@ class ContestController extends Controller {
         $accesses = ProblemsetAccessLogDAO::GetAccessForProblemset($response['contest']->problemset_id);
         $submissions = SubmissionLogDAO::GetSubmissionsForProblemset($response['contest']->problemset_id);
 
-        return ActivityReport::getActivityReport($accesses, $submissions);
+        return [
+            'status' => 'ok',
+            'events' => \OmegaUp\ActivityReport::getActivityReport($accesses, $submissions),
+        ];
     }
 
     /**
